# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_opsmx_ai-guardian-api_pkg_service.CreateHubRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      tag:
        type: string
    required:
    - email
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: AI Guardian API Service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: AI Guardian API
  version: "1.0"
paths:
  /api/v1/audit/report:
    get:
      consumes:
      - application/json
      description: Returns the users audit report
      parameters:
      - description: From timestamp/date filter
        in: query
        name: from
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No report found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get users audit report
      tags:
      - Audit
...
  /api/v1/hubs:
    post:
      consumes:
      - application/json
      description: Creates a new hub with the provided details
      parameters:
      - description: Hub creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_opsmx_ai-guardian-api_pkg_service.CreateHubRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new hub
      tags:
      - Hubs
...
  /api/v1/hubs/{id}:
    get:
      consumes:
      - application/json
      description: Returns the hub with the specified ID
      parameters:
      - description: Hub ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid hub ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Hub not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get hub by ID
      tags:
      - Hubs
  /api/v1/hubs/{id}/stats:
    get:
      consumes:
      - application/json
      description: Returns the hub stats with the specified ID
      parameters:
      - description: Hub ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid hub ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Hub not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get hub stats by ID
      tags:
      - Hubs
  /api/v1/hubs/user/list:
    get:
      consumes:
      - application/json
      description: Returns a list of hubs owned by the specified user
      parameters:
      - description: Email of the hub owner
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing or invalid email parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: List hubs by owner
      tags:
      - Hubs
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"